flask/requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
DateTime==4.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.2
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
zope.interface==5.4.0
